Added the beginning functionality of web site crawling
I have added the configuration and general basics for putting together a complete crawl of a website.  I have not put in the actual lists for processed and processing and the crawl, but I do have the looping ready and the process started.  Tomorrow I will add those peices to make it completely functional.