Added a max page crawl
I have added the max page crawl to the website source.  This will allow you to make sure that you hold to a managable size for travering the website.  I have also added "####" between the html for each page in the webData variable.  Thus, we can not determine the begining and end of each page stored in the variable.  Now I will have to add a backoff time and modify the config file reader to parse the new website variables